Created a new process image entry point
